fixed warnings on linux, restructure folders

+++ lib/util/util.cpp
@@ -11,8 +11,8 @@ using std::endl;
 Util::Util() { };
 
 void Util::print_matrix(float* matrix, size_t m, size_t n) {
-	for(int i = 0; i < m; i++) {
-		for(int j = 0; j < n; j++) {
+	for(size_t i = 0; i < m; i++) {
+		for(size_t j = 0; j < n; j++) {
 			cout << matrix[i * m + j] << "\t";
 		}
 		cout << endl;
